+++ lib/vagrant/config/vagrant.rb
@@ -4,15 +4,10 @@ module Vagrant
       configures :vagrant
 
       attr_accessor :dotfile_name
-      attr_accessor :home
       attr_accessor :host
 
-      def initialize
-        @home = nil
-      end
-
       def validate(errors)
-        [:dotfile_name, :home, :host].each do |field|
+        [:dotfile_name, :host].each do |field|
           errors.add(I18n.t("vagrant.config.common.error_empty", :field => field)) if !instance_variable_get("@#{field}".to_sym)
         end
       end